move modals to kaoto toolbar
consolidate pages + layout dirs

move kaoto toolbar back to components
