+++ src/components/Catalog.test.tsx
@@ -1,5 +1,5 @@
+import { AlertProvider } from '../layout/MASAlerts';
 import { Catalog } from './Catalog';
-import { AlertProvider } from './MASAlerts';
 import { screen } from '@testing-library/dom';
 import { render } from '@testing-library/react';
