+++ src/components/index.ts
@@ -2,13 +2,10 @@ export * from './Catalog';
 export * from './ConfirmationModal';
 export * from './Console';
 export * from './DeploymentsModal';
-export * from './ErrorBoundary';
 export * from './Extension';
 export * from './JsonSchemaConfigurator';
 export * from './KaotoDrawer';
 export * from './KaotoToolbar';
-export * from './MASAlerts';
-export * from './MASLoading';
 export * from './MiniCatalog';
 export * from './SettingsModal';
 export * from './SourceCodeEditor';
